% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{nullrangesData}
\alias{nullrangesData}
\title{ExperimentHub datasets for the nullranges package}
\description{
DNase hypersensitivity sites (DHS), CTCF binding sites, and
CTCF genomic interactions for demonstration of functions in the
nullranges package.
}
\examples{

suppressPackageStartupMessages(library(GenomicRanges))
dhs <- DHSA549Hg38
dhs

}
